requirements.txt
# Required Modules
beautifulsoup4
flask
python-dotenv
requests
# Dev Modules
freezegun
